# [ROCm] Add FP8 rowwise support to _scaled_grouped_mm + Submodule update