@cthi 2025-07-24
@pytorchbot label "module: rocm, release notes: rocm"